requirements.txt
google-generativeai==0.8.0
fastapi
uvicorn
torch
torchvision
pillow
python-multipart